# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CDED Clocking Device API",
    "description": "API Documentation for CDED'S Clocking Device",
    "version": "1.0.0"
  },
  "paths": {
    "/api/clocking/{uuid}/toggle-clock": {
      "get": {
        "summary": "Toggle clocked-in status of a card",
        "description": "Endpoint to toggle the clocked-in status of a card identified by its unique Card_UUID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "The unique identifier of the card to toggle the clocked-in status.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isClockedIn": {
                      "type": "boolean",
                      "description": "The updated clocked-in status of the card."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clocking": {
      "get": {
        "summary": "View all registered cards",
        "description": "Endpoint to retrieve details of all registered cards.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clocking/{Card_UUID}": {
      "get": {
        "summary": "View details of a registered card",
        "description": "Endpoint to retrieve details of a specific registered card identified by its unique Card_UUID.",
        "parameters": [
          {
            "name": "Card_UUID",
            "in": "path",
            "description": "The unique identifier of the registered card to retrieve.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          }
        }
      }
    },
    "/api/clocking/delete/{Card_UUID}": {
      "delete": {
        "summary": "Delete a registered card",
        "description": "Endpoint to delete details of a specific registered card identified by its unique Card_UUID.",
        "parameters": [
          {
            "name": "Card_UUID",
            "in": "path",
            "description": "The unique identifier of the registered card to delete.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "A message indicating the status of the operation."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clocking/register": {
      "post": {
        "summary": "Register a new card",
        "description": "Endpoint to register a new card.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCard"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          }
        }
      }
    },
    "/api/clocking/update/{Card_UUID}": {
      "put": {
        "summary": "Update details of a registered card",
        "description": "Endpoint to update details of a registered card identified by its unique Card_UUID.",
        "parameters": [
          {
            "name": "Card_UUID",
            "in": "path",
            "description": "The unique identifier of the registered card to update.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Card"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Card": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32",
            "description": "The unique identifier of the card."
          },
          "Card_UUID": {
            "type": "string",
            "description": "The UUID of the card."
          },
          "First_Name": {
            "type": "string",
            "description": "The first name of the card holder."
          },
          "Middle_Name": {
            "type": "string",
            "description": "The middle name of the card holder."
          },
          "Last_Name": {
            "type": "string",
            "description": "The last name of the card holder."
          },
          "Reg_No": {
            "type": "string",
            "description": "The registration number of the card."
          },
          "Department": {
            "type": "string",
            "description": "The registration number of the card."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the card was created."
          }
        }
      },
      "NewCard": {
        "type": "object",
        "properties": {
          "Card_UUID": {
            "type": "string",
            "description": "The UUID of the card."
          },
          "First_Name": {
            "type": "string",
            "description": "The first name of the card holder."
          },
          "Middle_Name": {
            "type": "string",
            "description": "The middle name of the card holder."
          },
          "Last_Name": {
            "type": "string",
            "description": "The last name of the card holder."
          },
          "Reg_No": {
            "type": "string",
            "description": "The registration number of the card."
          },
          "Department": {
            "type": "string",
            "description": "The department of the card."
          }
        }
      }
    }
  }
}